Ensure TekuCLIException is also handled on TekuDefaultExceptionHandler
