+++ infrastructure/exceptions/src/main/java/tech/pegasys/teku/infrastructure/exceptions/TekuCLIException.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright Consensys Software Inc., 2023
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.infrastructure.exceptions;
+
+public class TekuCLIException extends RuntimeException {
+  private final int resultCode;
+
+  public TekuCLIException(int resultCode) {
+    super("Unable to start Teku. Exit code: " + resultCode);
+    this.resultCode = resultCode;
+  }
+
+  public int getResultCode() {
+    return resultCode;
+  }
+}
